Distinguish between active AmigaOS and ROM version
Starting with AmigaOS 3.9, it is possible that the ROM version is
completely different to the active AmigaOS version. To take this into
account, IDHW_OSVER now returns the version of the active AmigaOS, which
is actually the expected value. The version of the physical ROM is
returned by the new IDHW_ROMVER query, and might be useful for bug
report purposes.